server/requirements.txt
numpy==1.16.4
Pillow-SIMD==6.0.0.post0
pyzmq==18.0.2
six==1.12.0
torch==1.1.0
torchvision==0.3.0
